Fix login gate bug, move model files to dedicated folder
