+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/models/mesh-browser-models.ts
@@ -11,7 +11,6 @@ export interface MeshSchemaListResponse {
     schemas: Array<Schema>
 }
 
-
 export interface Schema {
     name: string;
     elements: Array<SchemaElement>;
