+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/providers/index.ts
@@ -1,3 +1,3 @@
 export * from './mesh-browser-loader.service'
 export * from './mesh-browser-repository-table-loader.service'
-export * from './mesh-browser-models'
+export * from '../models/mesh-browser-models'
